web/backend/requirements.txt
fastapi>=0.116.0
uvicorn>=0.35.0
python-dotenv>=1.1.0
sqlalchemy>=2.0.0
pymysql>=1.1.1
